React: create helper for storing state in URL params
react-router-dom comes with a useSearchParams() hook which allows
all of the GET request parameters to be accessed for the current
URL, however this adds in a lot of mess to access a specific value.

This creates urlHelpers.js with a useSearchParam() function to
allow for cleaner access of a single parameter, including a
default value.

AllGames' textQuery has been moved over to this as an example.