All Players/Games: fix back button
Previously, every change to url state would be saved into the
browser history, causing undesirable behaviour in the back button
such as undoing the text query 1 character at a time before
actually moving to the previous page.

urlHelpers has been modified to allow for options to be passed
down to react-router, and these pages now enable the 'replace'
option so that only the current history entry is modified in-place.